pyproject.toml
[tool.poetry]
name = "crewai-azure"
version = "0.1.0"
description = "An AI-powered framework utilizing Azure OpenAI services via CrewAI."
authors = ["Rishikesh Gupta <therishiyon@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "^0.11.0"
python-dotenv = "^1.0.0"
langchain_openai = "^0.0.2"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.9.0"
ruff = "^0.1.0"  # Valid version
mypy = "^1.5.1"
pytest-cov = "^4.1.0"

[tool.pyright]
useLibraryCodeForTypes = true
exclude = [".cache", ".venv", "__pycache__"]

[tool.ruff]
select = ["E", "W", "F", "I", "B", "C4", "ARG", "SIM"]
ignore = ["W291", "W292", "W293"]
fixable = ["E", "F", "B"]
exclude = ["migrations", "tests/fixtures"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
